requirements.txt
# FERREMAS - Dependencias del Proyecto
# E-commerce para productos ferreteros

# === FRAMEWORK WEB ===
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Mail==0.10.0
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.4

# === BASE DE DATOS ===
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31

# === SERIALIZACIÓN Y VALIDACIÓN ===
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0

# === DOCUMENTACIÓN API ===
flasgger==0.9.7.1

# === PAGOS - WEBPAY PLUS ===
transbank-sdk==4.0.0

# === REQUESTS Y API EXTERNA ===
requests==2.32.3
urllib3==2.2.2

# === VARIABLES DE ENTORNO ===
python-dotenv==1.0.1

# === UTILIDADES ===
python-dateutil==2.9.0
pytz==2024.1

# === TESTING ===
pytest==8.4.1
pytest-cov==5.0.0
pytest-mock==3.14.1
pytest-flask==1.3.0
pytest-html==4.1.1

# === DESARROLLO ===
black==24.8.0
flake8==7.1.1
isort==5.13.2

# === PRODUCCIÓN ===
gunicorn==23.0.0
supervisor==4.2.5

# === SEGURIDAD ===
cryptography==43.0.0
bcrypt==4.2.0

# === LOGGING Y MONITOREO ===
structlog==24.4.0

# === CACHE (OPCIONAL) ===
redis==5.0.8
Flask-Caching==2.3.0 